// src/app/basic-setup/models/CourseSection.ts
export interface CourseSection {
    courseSectionId: number;
    baseSchoolNameId:number;
    BnaCurriculumTypeId:number;
   // curryculumType: string;
    courseNameId:number;
    sectionName: string;
    courseName: string;
    sectionShortName: string;
    menuPosition: number;
    isActive: boolean;
   
}